+++ dynamo_consistency/consistency_config.json
@@ -10,30 +10,22 @@
   "CacheLocation": "cache",
   "LogLocation": "logs",
   "Redirectors": {
-    "T3_CH_PSI": "t3se01.psi.ch",
     "T3_US_MIT": "t3serv006.mit.edu"
   },
-  "PathPrefix": {
-    "T2_UK_London_Brunel": "/cms"
-  },
-  "AccessMethod": {
-    "T2_FR_GRIF_LLR" : "SRM"
-  }, 
+  "PathPrefix": {},
+  "AccessMethod": {},
+  "LFNRoot": "/store",
   "DirectoryList": [
     "mc",
     "data",
     "hidata"
   ],
-  "IgnoreDirectories": [
-  ],
+  "IgnoreDirectories": [],
   "GlobalRedirectors": [
-    "xrootd-redic.pi.infn.it",
-    "cmsxrootd1.fnal.gov"
+    "cms-xrd-global.cern.ch"
   ],
   "SaveCache": 0,
-  "UseLoadBalancer": [
-    "T2_US_Vanderbilt"
-  ],
+  "UseLoadBalancer": [],
   "WebDir": "www",
   "MaxMissing": 100,
   "MaxOrphan": 1000,
